[CHANGE]: pay rate to float

+++ db/migrate/20190627141823_create_employees.rb
@@ -3,7 +3,7 @@ class CreateEmployees < ActiveRecord::Migration[5.2]
     create_table :employees do |t|
       t.string :full_name
       t.string :pay_type
-      t.integer :pay_rate
+      t.float :pay_rate
       t.string :active_status, default: 'Active'
       t.string :filing_status
       t.integer :w4_allowance, default: 0
